Unified Script tests
